db integration
