+++ client/pages/register.html
@@ -142,7 +142,7 @@
   <!-- Template Main JS File -->
   <script src="../assets/js/main.js"></script>
 
-  <script src="../scripts/script.js"></script>
+  <script src="../scripts/register.js"></script>
 
 </body>
