+++ server/.env
@@ -1,4 +1,4 @@
 PORT = 5000
-LOCAL_URI = "mongodb://127.0.0.1:27017/admin"
+# MONGO_URI = "mongodb://127.0.0.1:27017/admin"
 MONGO_URI = "mongodb+srv://nyabisisara:2SbOEXqo4k1Cgv8b@cluster0.jqcjcbd.mongodb.net/chama?retryWrites=true&w=majority"
 SECRET = 'a4cc42bbad6b239d90437eba3fa2e7cd085037f21b2252e85becacc7f0fa106d5df11f8401b4bac052640fd2423a0503124cd3d82bf1f886a3491be802551121'
