+++ server/index.js
@@ -15,4 +15,5 @@ const connection = async() => {
     }
 }
 
+
 connection() && server.listen(PORT, () => console.log(`Listening for requests on port ${port}`) )
