+++ server/req.rest
@@ -1,9 +1,9 @@
-GET http://localhost:5000/users
-# Content-Type: application/json
+POST http://localhost:5000/signup
+Content-Type: application/json
 
-# {
-#     "name": "kassim ali",
-#     "email": "kassim@yahoo.com",
-#     "username": "caseem",
-#     "password": "kassimali2001"
-# }
+{
+    "name": "kassim ali",
+    "email": "kassim@yahoo.com",
+    "username": "caseem",
+    "password": "kassimali2001"
+}
